+++ app/controllers/meals_controller.rb
@@ -71,11 +71,16 @@ class MealsController < ApplicationController
   end
 
   def delete_all
-    if current_user.meals.delete_all
-      redirect_to user_meals_path(current_user), notice: 'All meals deleted successfully'
+    if current_user.meals.any?
+      if current_user.meals.delete_all
+        redirect_to user_meals_path(current_user), notice: 'All meals deleted successfully'
+      else
+        redirect_to user_meals_path(current_user),
+          :flash => { :error => "Something went wrong while deleting meals" }
+      end
     else
       redirect_to user_meals_path(current_user),
-        :flash => { :error => "Something went wrong while deleting meals" }
+        :flash => { :error => "No meals to delete" }
     end
   end
