+++ app/controllers/routines_controller.rb
@@ -86,11 +86,16 @@ class RoutinesController < ApplicationController
   end
 
   def delete_all
-    if current_user.routines.delete_all
-      redirect_to user_routines_path(current_user), notice: 'All routines deleted successfully'
+    if current_user.routines.any?
+      if current_user.routines.delete_all
+        redirect_to user_routines_path(current_user), notice: 'All routines deleted successfully'
+      else
+        redirect_to user_routines_path(current_user),
+          :flash => { :error => "Something went wrong while deleting routines" }
+      end
     else
       redirect_to user_routines_path(current_user),
-        :flash => { :error => "Something went wrong while deleting routines" }
+        :flash => { :error => "No routines to delete" }
     end
   end
